I think a gecko is a type of lizard, right? Either way, this is the LIZARD UPDATE (initial bugfixes and changed some behavior)
1) Added several lines of code related to bugs with inconsistent
formatting / display / logging of information, and made the verbose
"debug" logging less... just less (much of information is now in
the default level without needing to enabled debug logging)... Also
changed the grammar on various things to prevent linewrapping
from too much words, or simple cleanups to make 8 decimal places be
usd correctly (in addition to fixing a few rounding errors)

2) Hardcoded a trading fee for cex.io platform, as it now has one (this
value doesn't seem to have a way to fetch via API, so I guess it can be
configurable in the future. The config parser needs an audit though, so
the workaround was to simply hardcode a non-zero fee)

3) Included the "launcher" script I personally use, as the provided one
doesn't restart itself automatically.

4) Removed the "sell when price is high" logic. This fork of gekkobot is
for reinvestment of the income from cloud mining. Period. Full stop.